Moved Login to Index
