+++ app/index.tsx
@@ -1,11 +1,10 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "expo-router";
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, TextInput, Button, Text, StyleSheet, TouchableOpacity, Image, ImageBackground } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import SplashScreen from "../components/SplashScreen";
 
 const HomePage: React.FC = () => {
-  const navigation = useNavigation();
   const [isSplashVisible, setSplashVisible] = useState(true);
 
   useEffect(() => {
@@ -20,56 +19,106 @@ const HomePage: React.FC = () => {
     return <SplashScreen />;
   }
 
+  const navigation = useNavigation();
+  const [login, setLogin] = useState({ username: '', password: '' });
+
+  const handleInput = (field, value) => {
+    setLogin((prevForm) => ({ ...prevForm, [field]: value }));
+  };
+
+  const handleLogin = () => {
+    navigation.navigate('mainscreen');
+    // Put backend auth here with implementing
+  };
+
+  const handleForgotPassword = () => {
+    // Todo
+  };
+
+
   return (
-    /*<View>
-        <Text>Home Page</Text>
-        <Link href="/weekplanscreen">
-          <Text>Gå til ugeplan</Text>
-        </Link>
-      </View>
-    );*/
-    <View style={styles.container}>
-      <Text style={styles.title}>Home Page</Text>
-      <Link href="/example" style={styles.link}>
-        <Text style={styles.link}>Gå til Example Page</Text>
-      </Link>
-      <Link href="loginscreen" style={styles.link}>
-        <TouchableOpacity
-          style={styles.button}
-          onPress={() => navigation.navigate('loginscreen')}
-        >
-          <Text style={styles.buttonText}>Gå til Login</Text>
+  <View style={styles.app}>
+  <ImageBackground source={require('../assets/backgroundImage.png')} resizeMode='cover' style={styles.app}>
+    <View style={styles.logincontainer}>
+      <Text style={styles.title}>Login</Text>
+      <TextInput
+        style={styles.input}
+        placeholder="Username"
+        value={login.username}
+        onChangeText={(value) => handleInput('username', value)}
+      />
+      <TextInput
+        style={styles.input}
+        placeholder="Password"
+        secureTextEntry
+        value={login.password}
+        onChangeText={(value) => handleInput('password', value)}
+      />
+      <View style={styles.buttonContainer}>
+        <Button title="Login" onPress={handleLogin} />
+        <TouchableOpacity onPress={handleForgotPassword}>
+          <Text style={styles.forgotPasswordText}>Forgot Password?</Text>
         </TouchableOpacity>
-      </Link>
+      </View>
     </View>
+      <Image
+        style={styles.image}
+        source={require('../assets/giraffe_ex1.png')}
+      />
+  </ImageBackground>
+  </View>
   );
 };
 
-// Styles for the HomePage component
 const styles = StyleSheet.create({
-  container: {
+  app: {
     flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
-    backgroundColor: '#A7C6ED',
+    backgroundColor: '#F9EBAE',
+    width: '100%',
+  },
+  logincontainer: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'white',
+    borderRadius: 10,
+    height: 300,
+    width: 250,
+    zIndex: 2,
   },
   title: {
     fontSize: 24,
     marginBottom: 20,
     fontWeight: 'bold',
   },
-  link: {
-    marginVertical: 10,
-  },
-  button: {
-    backgroundColor: '#B7E2D8',
-    padding: 15,
+  input: {
+    width: '80%',
+    padding: 10,
+    marginBottom: 15,
+    backgroundColor: '#fff',
     borderRadius: 5,
+    borderWidth: 1,
+    borderColor: '#ccc',
   },
-  buttonText: {
-    color: '#4A4A4A',
-    textAlign: 'center',
-    fontSize: 18,
+  buttonContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    width: '100%',
+    marginTop: 10,
+  },
+  forgotPasswordText: {
+    color: '#000000',
+    marginLeft: 10,
+  },
+  image: {
+    position: 'absolute',
+    alignSelf: 'flex-end',
+    width: 200,
+    resizeMode: 'contain',
+    zIndex: 1,
+    right: -20,
   },
 });
